+++ Calculator.Matija87/Calculator.Matija87.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34003.232
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator.Matija87", "Calculator.Matija87\Calculator.Matija87.csproj", "{B641F701-89FC-4A0C-A24E-BDF06093A1F0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Calculator.Matija87", "Calculator.Matija87\Calculator.Matija87.csproj", "{B641F701-89FC-4A0C-A24E-BDF06093A1F0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorLibary", "CalculatorLibary\CalculatorLibary.csproj", "{AE63B283-CE88-47FA-A8C1-51665A10B0FB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{B641F701-89FC-4A0C-A24E-BDF06093A1F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B641F701-89FC-4A0C-A24E-BDF06093A1F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B641F701-89FC-4A0C-A24E-BDF06093A1F0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE63B283-CE88-47FA-A8C1-51665A10B0FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE63B283-CE88-47FA-A8C1-51665A10B0FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE63B283-CE88-47FA-A8C1-51665A10B0FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE63B283-CE88-47FA-A8C1-51665A10B0FB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
