+++ Calculator.Matija87/CalculatorLibary/CalculatorLibary.csproj
@@ -1,14 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\CalculatorLibary\CalculatorLibary.csproj" />
-  </ItemGroup>
 
 </Project>
